[mongodb] * Moved routing methods to router class * Added mongodb docker compose yaml file
