+++ docker-compose.yml
@@ -1,0 +1,7 @@
+version: '3'
+services:
+    mongo:
+        container_name: mongo
+        image: mongo
+        ports:
+            - '27017:27017'
